[SPARK-57590][SQL] Address review: trim comments, inline localizeEntries, share localize-path tests
- Inline the trait localizeEntries delegate into readLocalizedEntries.
- Simplify readLocalizedEntries hasNext to !done (advance() ensures the invariant).
- Brace the object localizeEntries if/else; drop stale comments in
  SupportsArchiveFormat and readArchiveFooters.
- ParquetFileFormat: drop the isSplitable comment and gate the wrapped-corrupt footer
  catch on isArchivePath so loose files are unchanged.
- Move the localize-path shared tests (differing-field read, corrupt-entry inference
  and read) into ArchiveReadSuiteBase behind a localizesEntries flag; drop Parquet's
  redundant encodeFile override.
