+++ sql/core/src/test/scala/org/apache/spark/sql/execution/datasources/ArchiveReadSuiteBase.scala
@@ -101,6 +101,10 @@ trait ArchiveReadSuiteBase extends QueryTest with SharedSparkSession {
   /** Writes bytes that are not a readable archive at `dest` (of [[corruptArchiveExtension]]). */
   protected def writeCorruptArchive(dest: File): Unit
 
+  /** Bytes that are not a valid data file of [[format]], used to plant a corrupt archive entry. */
+  protected def corruptEntryBytes: Array[Byte] =
+    s"This is not a valid $format file".getBytes("UTF-8")
+
   /** An archive extension whose reader fails on corrupt bytes (used by the corrupt-file tests). */
   protected def corruptArchiveExtension: String
 
@@ -157,6 +161,14 @@ trait ArchiveReadSuiteBase extends QueryTest with SharedSparkSession {
    * to infer instead of falling through to a readable one. The streaming formats scan all inputs.
    */
   protected def inferenceSamplesOneFile: Boolean = false
+
+  /**
+   * Whether this format unpacks each archive entry to a local temp file before reading it (Parquet
+   * needs random access), as opposed to streaming entries. Gates the shared tests that exercise the
+   * localize path: per-entry corrupt handling during inference and read, and the differing-field
+   * read.
+   */
+  protected def localizesEntries: Boolean = false
 
   /** Sample data with a nested struct column, used by the complex-type test. */
   protected def complexSampleDf: DataFrame =
@@ -436,6 +448,72 @@ trait ArchiveReadSuiteBase extends QueryTest with SharedSparkSession {
           Files.write(new File(looseDir, s"loose.$fileExtension").toPath, encodeFile(loose))
           assert(schema == inferredSchema(Seq(looseDir.getCanonicalPath)),
             s"differing-field inference diverged from a directory read; got $schema")
+        }
+      }
+    }
+  }
+
+  // ----- shared localize-path tests (run when `localizesEntries`) ------------
+
+  if (localizesEntries) {
+    // The merge-gated block above owns this test when a format unions by name; localize-path
+    // formats (Parquet) set `supportsSchemaMerge` off, so provide it here instead.
+    if (!supportsSchemaMerge) {
+      test("archive entries with differing fields read like a directory") {
+        // One entry carries an extra field the other lacks; read under a schema covering both, the
+        // missing field reads back null -- exactly as a directory read of the same files does.
+        val withName = sampleDf((1, "Alice"), (2, "Bob"))
+        val idOnly = Seq(3).toDF("id")
+        assertArchiveMatchesDir(
+          Seq(entryName(0) -> encodeFile(withName), entryName(1) -> encodeFile(idOnly)))
+      }
+    }
+
+    Seq(true, false).foreach { ignoreCorrupt =>
+      test(s"ignoreCorruptFiles=$ignoreCorrupt: inference skips a corrupt entry's whole archive") {
+        // A corrupt entry condemns its whole archive during inference (no partial ingestion), so
+        // the `extra` column carried by the bad archive's valid sibling entry must not surface. A
+        // good archive is unaffected. Formats sampling one file need mergeSchema to fold every
+        // entry's footer.
+        val extra =
+          if (inferenceSamplesOneFile) Map("mergeSchema" -> "true") else Map.empty[String, String]
+        val sibling = Seq((3, "Carol", "x")).toDF("id", "name", "extra")
+        withTempDir { dir =>
+          writeArchive(new File(dir, s"good.${archiveExtensions.head}"),
+            Seq(entryName(0) -> encodeFile(sampleDf((1, "Alice"), (2, "Bob")))))
+          writeArchive(new File(dir, s"bad.${archiveExtensions.head}"), Seq(
+            entryName(0) -> encodeFile(sibling),
+            entryName(1) -> corruptEntryBytes))
+          withSQLConf(SQLConf.IGNORE_CORRUPT_FILES.key -> ignoreCorrupt.toString) {
+            if (ignoreCorrupt) {
+              assert(inferredSchema(Seq(dir.getCanonicalPath), extra).fieldNames.toSet ==
+                Set("id", "name"),
+                "the whole corrupt archive, including its valid sibling entry, should be skipped")
+            } else {
+              intercept[Exception](inferredSchema(Seq(dir.getCanonicalPath), extra))
+            }
+          }
+        }
+      }
+
+      test(s"ignoreCorruptFiles=$ignoreCorrupt: read skips the rest of an archive at a bad entry") {
+        // The read ingests entries before the corrupt one, then skips the rest of that archive (not
+        // whole-archive atomic like inference). A separate good archive is read in full.
+        val alive = sampleDf((1, "Alice"), (2, "Bob"))
+        val beforeCorrupt = sampleDf((3, "Carol"))
+        withTempDir { dir =>
+          writeArchive(new File(dir, s"good.${archiveExtensions.head}"),
+            Seq(entryName(0) -> encodeFile(alive)))
+          writeArchive(new File(dir, s"bad.${archiveExtensions.head}"), Seq(
+            entryName(0) -> encodeFile(beforeCorrupt),
+            entryName(1) -> corruptEntryBytes))
+          withSQLConf(SQLConf.IGNORE_CORRUPT_FILES.key -> ignoreCorrupt.toString) {
+            if (ignoreCorrupt) {
+              checkAnswer(read(dir.getCanonicalPath), alive.union(beforeCorrupt))
+            } else {
+              intercept[SparkException](read(dir.getCanonicalPath).collect())
+            }
+          }
         }
       }
     }
